backend/requirements.txt
Flask==1.1.2
Flask-AWSCognito==1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.22